- GetApplicationStatus method added to ClsApplication business logic - In ClsDataType layer New application status value added to EnApplicationStatus which is "InProgress" to help enable and disable Tests Tool Menu Strip in manage local application - _EnableDisableScheduleTestsToolStripsMenu method added in manage local applications form

+++ DVLD.DataAccess/ClsApplications.cs
@@ -341,7 +341,7 @@ namespace DVLD.DataAccess
                         object result = command.ExecuteScalar();
                         if (result != null)
                         {
-                            ApplicationStatus = (int)result;
+                            ApplicationStatus = (byte)result;
                         }
                     }
                     catch (Exception ex)
@@ -353,8 +353,6 @@ namespace DVLD.DataAccess
 
             return (EnApplicationStatus)ApplicationStatus;
         }
-       
-
 
     }
 }
